feat: 장바구니 조회 응답 필드 확장 및 금액 타입 정리

+++ team1/src/main/java/com/gdg/sprint/team1/dto/cart/CartSummary.java
@@ -1,9 +1,11 @@
 package com.gdg.sprint.team1.dto.cart;
 
+import java.math.BigDecimal;
+
 public record CartSummary (
         Integer totalItems,
         Integer totalQuantity,
-        Integer totalProductPrice,
-        Integer deliveryFee,
-        Integer finalPrice
+        BigDecimal totalProductPrice,
+        BigDecimal deliveryFee,
+        BigDecimal finalPrice
 ) {}
